Dev: moved useGasSummary back to where it was
